requirements (1).txt
numpy
matplotlib
gradio==3.41.2
transformers
Pillow
scikit-learn
sentencepiece
torch
